requirements.txt
streamlit
google-generativeai>=0.5.0
requests
python-dotenv
